Deleted Service files

+++ src/main/java/com/codecool/flight_api_project/airline/Airline.java
@@ -9,7 +9,7 @@ public class Airline
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String name;
 
 
@@ -24,7 +24,7 @@ public class Airline
     {
     }
 
-    public int getId()
+    public Long getId()
     {
         return id;
     }
